+++ examples/resources/nxos_system/resource.tf
@@ -129,4 +129,6 @@ resource "nxos_system" "example" {
       vrf_minimum                         = 4
     }
   }
+  smart_licensing_transport_mode     = "transportSmart"
+  smart_licensing_transport_cslu_url = "https://cslu.example.com"
 }