+++ internal/provider/resource_nxos_system_test.go
@@ -122,6 +122,8 @@ func TestAccNxosSystem(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "vdcs.1.vlan_minimum", "16"))
 	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "vdcs.1.vrf_maximum", "4096"))
 	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "vdcs.1.vrf_minimum", "4"))
+	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "smart_licensing_transport_mode", "transportSmart"))
+	checks = append(checks, resource.TestCheckResourceAttr("nxos_system.test", "smart_licensing_transport_cslu_url", "https://cslu.example.com"))
 	var tfVersion *goversion.Version
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
@@ -273,6 +275,8 @@ func testAccNxosSystemConfig_all() string {
 	config += `			vrf_minimum = 4` + "\n"
 	config += `		}` + "\n"
 	config += `	}` + "\n"
+	config += `	smart_licensing_transport_mode = "transportSmart"` + "\n"
+	config += `	smart_licensing_transport_cslu_url = "https://cslu.example.com"` + "\n"
 	config += `}` + "\n"
 	return config
 }